dashfrog_python_sdk/pyproject.toml
[project]
name = "dashfrog-python-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bandit>=1.8.6",
    "celery>=5.5.3",
    "celery-types>=0.23.0",
    "click>=8.0.0",
    "psycopg[binary]>=3.2.0",
    "fastapi>=0.118.0",
    "fastapi-cli>=0.0.13",
    "flask>=3.1.2",
    "flask-cli>=0.4.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-jaeger>=1.11.1",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-instrumentation-aws-lambda>=0.58b0",
    "opentelemetry-instrumentation-botocore>=0.58b0",
    "opentelemetry-instrumentation-celery>=0.58b0",
    "opentelemetry-instrumentation-click>=0.58b0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-flask>=0.58b0",
    "opentelemetry-instrumentation-httpx>=0.58b0",
    "opentelemetry-instrumentation-openai>=0.47.3",
    "opentelemetry-instrumentation-pymongo>=0.58b0",
    "opentelemetry-instrumentation-requests>=0.58b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
    "opentelemetry-semantic-conventions>=0.58b0",
    "pdoc3>=0.11.6",
    "pydantic>=2.11.9",
    "pydantic-settings[yaml]>=2.11.0",
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "redis>=5.2.2",
    "ruff>=0.13.2",
    "setuptools>=80.9.0",
    "shortuuid>=1.0.13",
    "sqlalchemy>=2.0.43",
    "wheel>=0.45.1",
]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    ".bzr",
    ".git",
    ".svn",
    ".hg",
    "typings",
    "src/test",
    "src/tests",
    "src/alembic",
    "examples",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort.sections]
opentel = ["opentelemetry*"]
tests=["pytest", "test*"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true
force-sort-within-sections = true
forced-separate = ["tests", "opentelemetry"]
known-local-folder = ["tests"]
known-first-party=["dashfrog_python_sdk*"]
section-order = [
  "future",
  "standard-library",
  "third-party",
  "opentel",
  "first-party",
  "local-folder",
  "tests",
]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", # asserts allowed in tests
]

[tool.ruff.lint.pycodestyle]
max-line-length = 120

[tool.pyright]
include = ["src/**"]
exclude = ["tests/**", ".venv/**", "examples/**"]
pythonVersion = "3.10"
stubPath = "typings"
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md#diagnostic-rule-defaults
deprecateTypingAliases = true
typeCheckingMode = "standard"

[tool.bandit]
exclude_dirs =  ["tests", ".venv", "examples"]
skips = ["B101"]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools]
packages = ["dashfrog_python_sdk"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
dashfrog_python_sdk = ["py.typed"]
